+++ Controls.h
@@ -31,7 +31,7 @@ void setup();
  * @param time
  * @param observer function to be invoked after checking for commands
  */
-void update(const uint16_t time, void(*observer)(COMMAND));
+void update(const uint32_t time, void(*observer)(COMMAND));
 
 // COMMAND_LIST getInput();
 };  // namespace controls
